Fixed some warnings on MSVS

+++ hide/Location.h
@@ -18,23 +18,23 @@ namespace hide
 	{
 	private:
     	std::string		_filename;
-        size_t			_line;
-        size_t			_column;
+        uint32_t		_line;
+        uint32_t		_column;
 
 	public:
 	    Location()
         	: _filename(), _line(0), _column(0)
         { }
 
-		Location(const std::string& filename, size_t line, size_t column)
+		Location(const std::string& filename, uint32_t line, uint32_t column)
 			: _filename(filename), _line(line), _column(column)
         { }
 
         bool IsValid() const { return !_filename.empty(); }
 
         std::string GetFilename() const	{ return _filename; }
-        size_t GetLine() const			{ return _line; }
-        size_t GetColumn() const		{ return _column; }
+        uint32_t GetLine() const		{ return _line; }
+        uint32_t GetColumn() const		{ return _column; }
 
 		HIDE_DECLARE_TO_STRING_METHOD()
 		HIDE_DECLARE_MEMBERS("filename", &Location::_filename, "line", &Location::_line, "column", &Location::_column)
